fix: swapping rank and world size reading to environment variables
For whatever reason, `torchrun` causes the rank and world size variables
to not be read correctly, even when `mpirun` does the correct thing.

The solution was to replace `dist.get_rank()` and `get_world_size` calls
from the `torch.distributed` API with OS environment variables. The uncertainty
with this, however, is how this will translate to multinode operations.